requirements.txt
# Web framework
fastapi
uvicorn

# Frontend
streamlit
plotly
pandas

# ML / AI
transformers
torch
scikit-learn
numpy
prophet

# Data sources
yfinance
newsapi-python
feedparser
requests
pytrends

# LLM
groq

# Environment
python-dotenv

# Misc
httpx